+++ web/app/(main-layout)/layout.tsx
@@ -7,7 +7,7 @@ import "material-icons/iconfont/material-icons.css";
 import CapacitorProvider from "@/components/providers/capacitor-provider";
 import RemoteModuleProvider from "@/components/providers/remote-module-provider";
 import InterModuleCommunicationProvider from "@/components/providers/imc-provider";
-import Nav from "@/components/interface/nav";
+import Nav from "@/components/interface/navigation/nav";
 import { Suspense } from "react";
 import { Analytics } from "@vercel/analytics/next";