+++ web/app/(main-layout)/page.tsx
@@ -3,11 +3,15 @@
 import CommandViewer from "@/components/interface/command-viewer";
 import EditorToolbar from "@/components/interface/editor-toolbar";
 import ExtensionSuggestionOverlay from "@/components/interface/extension-suggestion-overlay";
+import { EditorContext } from "@/components/providers/editor-context-provider";
 import Voice from "@/components/tools/voice";
 import ConsolePanelView from "@/components/views/console-panel/console-panel-view";
 import ViewArea from "@/components/views/view-area";
+import { useContext } from "react";
 
 export default function HomePage() {
+  const editorContext = useContext(EditorContext);
+
   return (
     <div className="relative flex h-full w-full flex-col">
       <EditorToolbar />
@@ -16,7 +20,7 @@ export default function HomePage() {
 
       <Voice />
       {false && <ExtensionSuggestionOverlay />}
-      <CommandViewer />
+      {editorContext?.editorStates.isCommandViewerOpen && <CommandViewer />}
     </div>
   );
 }