+++ web/components/interface/editor-toolbar.tsx
@@ -1,163 +1,23 @@
 "use client";
 
 import { Button, Divider, Tooltip } from "@heroui/react";
-import { useContext, useEffect, useState } from "react";
+import { useContext, useState } from "react";
 import Icon from "@/components/misc/icon";
 import AppSettingsModal from "@/components/modals/app-settings-modal";
 import { AnimatePresence, motion } from "framer-motion";
 import { EditorContext } from "../providers/editor-context-provider";
-import { getPlatform } from "@/lib/platform-api/platform-checker";
-import { PlatformEnum } from "@/lib/types";
-import toast from "react-hot-toast";
-import ExtensionMarketplaceModal from "../modals/extension-marketplace-modal";
 import AgentConfigModal from "../modals/agent-config-modal";
-import useSpeech2Speech from "@/lib/hooks/use-speech2speech";
-import { getAPIKey } from "@/lib/settings/api-manager-utils";
-import { editorAssistantAgent } from "@/lib/agent/built-in-agents/editor-assistant";
-import { getAgentLLMConfig, runAgentMethod } from "@/lib/agent/agent-runner";
-import useCommands from "@/lib/hooks/use-commands";
-import useTTS from "@/lib/hooks/use-tts";
-import { usePlatformApi } from "@/lib/hooks/use-platform-api";
-import { useTabViewManager } from "@/lib/hooks/use-tab-view-manager";
-import { getDefaultLLMConfig } from "@/lib/modalities/utils";
+import ExtensionMarketplaceModal from "../modals/extension-marketplace-modal";
+import usePlatformAIAssistant from "@/lib/hooks/use-platform-ai-assistant";
 
 export default function EditorToolbar() {
   const editorContext = useContext(EditorContext);
-  const { platformApi } = usePlatformApi();
+
+  const { runAssistant } = usePlatformAIAssistant();
 
   const [isAgentListModalOpen, setIsAgentListModalOpen] = useState(false);
   const [isExtensionModalOpen, setIsExtensionModalOpen] = useState(false);
   const [isAppSettingsModalOpen, setAppIsSettingsModalOpen] = useState(false);
-
-  const { runSpeech2Speech, stopSpeech2Speech, isUsing } = useSpeech2Speech();
-  const { readText, playAudio } = useTTS();
-  const { runCommand, commands: commandsInTab } = useCommands();
-
-  const [userVoiceMessage, setUserVoiceMessage] = useState<string>("");
-
-  const [assistantResult, setAssistantResult] = useState<any>(null);
-  const [pendingAnalysis, setPendingAnalysis] = useState("");
-  const [analysisAudio, setAnalysisAudio] = useState<Blob | undefined>(
-    undefined,
-  );
-
-  const { activeTabView } = useTabViewManager();
-
-  // When speech2speech returns assistant result, we let the assistant agent to analyze its effects
-  useEffect(() => {
-    async function processAssistantResult() {
-      if (!assistantResult) {
-        return;
-      }
-
-      const {
-        suggestedCmd,
-        suggestedArgs,
-        suggestedViewId,
-        response,
-      }: {
-        suggestedCmd: string;
-        suggestedArgs: {
-          name: string;
-          value: any;
-        }[];
-        suggestedViewId: string;
-        response: string;
-      } = assistantResult;
-
-      // TODO: The agent needs to confirm the command with the user
-      // TODO: before executing it.
-      const args = suggestedArgs.reduce(
-        (acc, arg) => {
-          acc[arg.name] = arg.value;
-          return acc;
-        },
-        {} as Record<string, any>,
-      );
-
-      // Show thinking indicator
-      editorContext?.setEditorStates((prev) => ({
-        ...prev,
-        isThinking: true,
-        thinkingText: "Executing command...",
-      }));
-
-      const cmdResult = await runCommand(suggestedViewId, suggestedCmd, args);
-
-      if (process.env.NODE_ENV === "development") {
-        console.log("Command result:", cmdResult);
-      }
-
-      if (!editorAssistantAgent.LLMConfig) {
-        toast.error("Agent is not configured to analyze command result.");
-        return;
-      }
-
-      const llmKey = getAPIKey(
-        editorContext,
-        editorContext?.persistSettings?.llmProvider,
-      );
-
-      if (!llmKey) {
-        toast.error("Please set your LLM API key in settings.");
-        return;
-      }
-
-      const { analysis }: { analysis: string } = await runAgentMethod(
-        llmKey,
-        editorAssistantAgent.LLMConfig,
-        editorAssistantAgent,
-        "analyzeCommandResult",
-        {
-          userMessage: userVoiceMessage,
-          suggestedCmd: suggestedCmd,
-          previousSuggestion: response,
-          commandResult: cmdResult,
-        },
-      );
-
-      if (process.env.NODE_ENV === "development") {
-        console.log("Command analysis:", analysis);
-      }
-      setPendingAnalysis(analysis);
-    }
-
-    processAssistantResult();
-  }, [assistantResult]);
-
-  // When the assistant agent is done analyzing the command result, we can
-  // play the analysis result to the user.
-  useEffect(() => {
-    if (pendingAnalysis.length > 0) {
-      if (!isUsing) {
-        // Show thinking indicator
-        editorContext?.setEditorStates((prev) => ({
-          ...prev,
-          isThinking: true,
-          thinkingText: "Analyzing command result...",
-        }));
-      }
-      readText(pendingAnalysis).then((blob) => {
-        setPendingAnalysis("");
-        setAnalysisAudio(blob);
-      });
-    }
-  }, [pendingAnalysis, isUsing]);
-
-  // Play the audio when the speech2speech is done and the analysis is done
-  useEffect(() => {
-    if (!isUsing && analysisAudio) {
-      // Show thinking indicator
-      editorContext?.setEditorStates((prev) => ({
-        ...prev,
-        isThinking: true,
-        thinkingText: "Analyzing command result...",
-      }));
-      playAudio(analysisAudio).then(() => {
-        setAnalysisAudio(undefined);
-      });
-    }
-  }, [isUsing, analysisAudio]);
 
   function setIsOpen(val: boolean) {
     if (editorContext) {
@@ -165,124 +25,6 @@ export default function EditorToolbar() {
         ...prev,
         isToolbarOpen: val,
       }));
-    }
-  }
-
-  async function handleMicFinished() {
-    if (!editorContext) {
-      return;
-    }
-
-    if (getPlatform() === PlatformEnum.VSCode) {
-      toast.error(
-        "Voice Chat is not supported in VSCode Extension. Please use other versions for Voice Chat.",
-      );
-      return;
-    }
-
-    if (!editorContext.editorStates.isRecording) {
-      const llmProvider = editorContext.persistSettings?.llmProvider;
-      const llmModel = editorContext.persistSettings?.llmModel;
-
-      if (!llmProvider || !llmModel) {
-        toast.error("Please set your LLM provider and model in settings.");
-        return;
-      }
-      const llmKey = getAPIKey(
-        editorContext,
-        editorContext.persistSettings?.llmProvider,
-      );
-
-      if (!llmKey) {
-        toast.error("Please set your LLM API key in settings.");
-        return;
-      }
-
-      const agent = editorAssistantAgent;
-      const methodName = "useExtensionCommands";
-
-      runSpeech2Speech(async (inputText: string) => {
-        const currentPath =
-          editorContext.persistSettings?.projectHomePath &&
-          editorContext.editorStates.project
-            ? editorContext.persistSettings?.projectHomePath +
-              "/" +
-              editorContext.editorStates.project
-            : undefined;
-
-        const projectDirTree = [];
-        if (currentPath) {
-          const gitIgnorePath = `${currentPath}/.gitignore`;
-          const gitIgnoreLines: string[] = [];
-          if (await platformApi?.hasPath(gitIgnorePath)) {
-            const gitIgnoreFile = await platformApi?.readFile(gitIgnorePath);
-            const gitIgnoreContent = await gitIgnoreFile?.text();
-            const lines = gitIgnoreContent?.split("\n") ?? [];
-            gitIgnoreLines.push(...lines.filter((line) => line.trim() !== ""));
-          }
-          const tree =
-            (await platformApi?.listPathContent(currentPath, {
-              include: "all",
-              gitignore: gitIgnoreLines,
-              isRecursive: true,
-            })) ?? [];
-          projectDirTree.push(...tree);
-        }
-
-        const config =
-          getAgentLLMConfig(agent, methodName) ??
-          getDefaultLLMConfig(editorContext);
-
-        if (!config) {
-          toast.error("No LLM config found for agent.");
-          return "No LLM config found for agent. Please configure the LLM in settings.";
-        }
-
-        setUserVoiceMessage(inputText);
-        // Pipe the LLM result to Speech2Speech
-        // const stream = await llm.generateStream(inputText);
-        const result = await runAgentMethod(llmKey, config, agent, methodName, {
-          chatHistory: [],
-          userMessage: inputText,
-          activeTabView,
-          commands: commandsInTab.map((commandInTab) => ({
-            cmd: commandInTab.commandInfo.name,
-            parameters: Object.entries(commandInTab.commandInfo.parameters).map(
-              ([key, value]) => ({
-                name: key,
-                type: value.type,
-                description: value.description,
-              }),
-            ),
-            description: commandInTab.commandInfo.description,
-            viewId: commandInTab.viewId,
-          })),
-          projectDirTree: projectDirTree,
-        });
-
-        const {
-          suggestedCmd,
-          suggestedArgs,
-          suggestedViewId,
-          response,
-        }: {
-          suggestedCmd: string;
-          suggestedArgs: {
-            name: string;
-            value: any;
-          }[];
-          suggestedViewId: string;
-          response: string;
-        } = result;
-
-        console.log("Agent suggestion:", result);
-
-        setAssistantResult(result);
-
-        return response;
-      });
-    } else {
-      stopSpeech2Speech();
     }
   }
 
@@ -364,7 +106,7 @@ export default function EditorToolbar() {
                   isIconOnly
                   className="text-default-foreground h-8 w-8 min-w-8 px-1 py-1"
                   onPress={() => {
-                    handleMicFinished();
+                    runAssistant(true);
                   }}
                   variant={
                     editorContext?.editorStates?.isRecording ? "solid" : "light"